server: Update billing plan operations

+++ migrations/20200123212019_init_billing.up.sql
@@ -4,7 +4,7 @@ CREATE TABLE billing_plans (
     updated_at TIMESTAMP WITH TIME ZONE NOT NULL,
 
     name TEXT NOT NULL,
-    price DOUBLE PRECISION NOT NULL,
+    price BIGINT NOT NULL,
     description TEXT NOT NULL,
     stripe_id TEXT NOT NULL,
     is_active BOOLEAN NOT NULL,
@@ -17,7 +17,7 @@ CREATE TABLE billing_plans (
 INSERT INTO billing_plans
 		(id, created_at, updated_at, name, description, stripe_id, price, is_active, tier, storage)
 		VALUES ('42fb1c42-caca-418d-81f3-a6313c4a0a42', '2020-02-10T14:33:39+00:00', '2020-02-10T14:33:39+00:00',
-    'Free', '', 'plan_Gck0Zy9Qx5qaGk', 0.0, true, 'FREE', 1000000000); -- 1GB
+    'Free', '', 'plan_Gck0Zy9Qx5qaGk', 0, true, 'FREE', 1000000000); -- 1GB
 
 CREATE TABLE billing_customers (
     id UUID NOT NULL,
